refactor(types): toggle
