+++ src/lib/builders/toggle/types.ts
@@ -1,0 +1,8 @@
+import type { createToggle } from './create';
+
+export type CreateToggleArgs = {
+	disabled?: boolean;
+	pressed?: boolean;
+};
+
+export type CreateToggleReturn = ReturnType<typeof createToggle>;
